Allow dynamically adding and removing threads from scheduler
Easy now that there isn't the ThreadPool intermediary

Reviewed-on: https://gerrit.dechocorp.com/9174

+++ mordor/iomanager_epoll.h
@@ -42,7 +42,7 @@ private:
     };
 
 public:
-    IOManager(int threads = 1, bool useCaller = true);
+    IOManager(size_t threads = 1, bool useCaller = true);
     ~IOManager();
 
     bool stopping();
