+++ mordor/iomanager_iocp.h
@@ -63,7 +63,7 @@ private:
     };
 
 public:
-    IOManager(int threads = 1, bool useCaller = true);
+    IOManager(size_t threads = 1, bool useCaller = true);
     ~IOManager();
 
     bool stopping();
