+++ mordor/iomanager_kqueue.h
@@ -39,7 +39,7 @@ private:
     };
 
 public:
-    IOManager(int threads = 1, bool useCaller = true);
+    IOManager(size_t threads = 1, bool useCaller = true);
     ~IOManager();
 
     bool stopping();
